rename plotter
