+++ lib/client/jsx/components/plotter/plot_series.jsx
