+++ lib/client/jsx/components/plotter/plot_variables.jsx
